+++ SmartCab.Backend/Api.DataAccessLayer/Repositories/CustomerRepository.cs
@@ -84,7 +84,7 @@ namespace Api.DataAccessLayer.Repositories
         /// /// <param name="deposit">The amount to deposit</param>
         /// <returns></returns>
         /// <exception cref="UserIdInvalidException">Customer does not exist.</exception>
-        public async Task DepositAsync(string customerId, int deposit)
+        public async Task DepositAsync(string customerId, decimal deposit)
         {
             var customer = await _context.Customers.FindAsync(customerId);
